requirements.txt
requests
pandas
joblib
google-cloud-bigquery
google-auth
pandas-gbq
pyyaml
dotenv
matplotlib